+++ src/libs/renderer/src/s_device.cpp
@@ -142,7 +142,7 @@ uint32_t DX9SetTexturePath(VS_STACK *pS)
     auto *pVR = (VDATA *)pS->Push();
     if (!pVR || iNumber < 0 || iNumber >= 4)
     {
-        pVR->Set(static_cast<int32_t>(0));
+        pVR->Set(0);
         return IFUNCRESULT_OK;
     }
 
@@ -150,7 +150,7 @@ uint32_t DX9SetTexturePath(VS_STACK *pS)
     DX9RENDER::pRS->TextureCreate((const char *)iNumber);
     DX9RENDER::pRS->TextureCreate(static_cast<const char *>(pStr));
 
-    pVR->Set(static_cast<int32_t>(1));
+    pVR->Set(1);
 
     return IFUNCRESULT_OK;
 }
@@ -164,7 +164,7 @@ uint32_t RPrint(VS_STACK *pS)
     if (pString->GetString())
         DX9RENDER::pRS->Print(x, y, pString->GetString());
     auto *pVR = (VDATA *)pS->Push();
-    pVR->Set(static_cast<int32_t>(0));
+    pVR->Set(0);
     return IFUNCRESULT_OK;
 }
 
@@ -177,7 +177,7 @@ uint32_t SetGlowParams(VS_STACK *pS)
     DX9RENDER::pRS->SetGLOWParams(fBlurBrushSize, Intensivity, BlurPasses);
 
     auto *pVR = (VDATA *)pS->Push();
-    pVR->Set(static_cast<int32_t>(0));
+    pVR->Set(0);
     return IFUNCRESULT_OK;
 }
